backend/requirements.txt
alembic==1.16.5
altair==5.5.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
bcrypt==4.0.1
blinker==1.9.0
cachetools==6.2.6
certifi==2026.2.25
charset-normalizer==3.4.5
click==8.1.8
dnspython==2.7.0
ecdsa==0.19.1
email-validator==2.3.0
exceptiongroup==1.3.1
fastapi==0.128.8
gitdb==4.0.12
GitPython==3.1.46
h11==0.16.0
idna==3.11
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
Mako==1.3.10
MarkupSafe==3.0.3
narwhals==2.17.0
numpy==2.0.2
packaging==25.0
pandas==2.3.3
passlib==1.7.4
pillow==11.3.0
protobuf==6.33.5
psycopg2-binary==2.9.11
pyarrow==21.0.0
pyasn1==0.6.2
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.20
pytz==2026.1.post1
referencing==0.36.2
requests==2.32.5
rpds-py==0.27.1
rsa==4.9.1
six==1.17.0
smmap==5.0.3
SQLAlchemy==2.0.48
starlette==0.49.3
streamlit==1.50.0
tenacity==9.1.2
toml==0.10.2
tomli==2.4.0
tornado==6.5.4
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uvicorn==0.39.0
